+++ docs/releases.md
@@ -4,6 +4,10 @@
 
 ### New Features
 
+- Support for sharded Zarr V3 arrays in `ZarrParser`.
+  ([#946](https://github.com/zarr-developers/VirtualiZarr/pull/946)).
+  By [Tom Nicholas](https://github.com/TomNicholas).
+
 ### Breaking changes
 
 ### Bug fixes
@@ -34,6 +38,9 @@ Brings `region`-writing support in `.to_icechunk()`, a `ZarrParser` with orders 
 
 ### Bug fixes
 
+- Fix handling of scalar Zarr V3 arrays with None dimension_names.
+  ([#897](https://github.com/zarr-developers/VirtualiZarr/pull/897)).
+  By [Lars Buntemeyer](https://github.com/larsbuntemeyer).
 - Fix setting `fill_value` for Zarr V2 arrays if data type is a subtype of integer or float.
   ([#845](https://github.com/zarr-developers/VirtualiZarr/pull/845)).
   By [Hauke Schulz](https://github.com/observingClouds).